feat: GraphQL-over-HTTP compliant error handling
