+++ endpoint/handler.go
@@ -1,20 +1,16 @@
 package endpoint
 
 import (
-	"errors"
+	"encoding/json"
 	"io"
 	"net/http"
+	"strings"
 	"sync"
 
-	"github.com/vektah/gqlparser/v2/ast"
-	"github.com/vektah/gqlparser/v2/parser"
 	"go.uber.org/zap"
 
 	"github.com/shinzonetwork/shinzo-network-gateway/host"
 )
-
-// ErrEmptyQuery is returned if GraphQL query is empty.
-var ErrEmptyQuery = errors.New("empty GraphQL query")
 
 // HostsSelector interface allows handler to get hosts for given collections.
 type HostsSelector interface {
@@ -46,19 +42,30 @@ func NewHandler(extractor CollectionsExtractor, selector HostsSelector, logger *
 
 func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	h.logger.Sugar().Debugw("serving HTTP request", "from", r.RemoteAddr)
+
+	mediaType := h.getContentType(r)
+	if mediaType == "" {
+		w.WriteHeader(http.StatusNotAcceptable)
+		return
+	}
+
 	body, err := io.ReadAll(r.Body)
 	if err != nil {
-		// TODO(tzdybal): handle errors
+		h.writeError(w, http.StatusBadRequest, err.Error(), mediaType)
+		return
 	}
 
 	collections, err := h.extractor.ExtractCollections(string(body))
 	if err != nil {
-		// TODO(tzdybal): handle errors
+		// GraphQL request errors: 400 for graphql-response+json, 200 for application/json per spec.
+		h.writeError(w, requestErrorStatus(mediaType), err.Error(), mediaType)
+		return
 	}
 
 	hosts, err := h.selector.SelectHosts(collections)
 	if err != nil {
-		// TODO(tzdybal): handle errors
+		h.writeError(w, http.StatusServiceUnavailable, err.Error(), mediaType)
+		return
 	}
 
 	responses := h.getHostsReponses(hosts, body)
@@ -87,31 +94,33 @@ func (h *Handler) composeResponse(w http.ResponseWriter, responses []hostRespons
 	panic("implement me!")
 }
 
-// CollectionsExtractor defines interface for extracting root collections from GraphQL queries.
-type CollectionsExtractor interface {
-	ExtractCollections(graphql string) ([]string, error)
+// getContentType picks the response content type from the Accept header.
+// As defined in GraphQL-over-HTTP spec: prefers application/graphql-response+json; falls back to application/json.
+// Returns "" if no supported type is acceptable (caller should respond 406).
+func (h *Handler) getContentType(r *http.Request) string {
+	accept := r.Header.Get("Accept")
+	switch {
+	case accept == "", strings.Contains(accept, "*/*"), strings.Contains(accept, contentTypeGraphQLResponse):
+		return contentTypeGraphQLResponse
+	case strings.Contains(accept, contentTypeJSON):
+		return contentTypeJSON
+	default:
+		return ""
+	}
 }
 
-// DefaultCollectionExtractor provides default implementation for root collections extraction.
-type DefaultCollectionExtractor struct{}
-
-// ExtractCollections parses GraphQL into AST and then traverse to get the root collections.
-func (e *DefaultCollectionExtractor) ExtractCollections(graphql string) ([]string, error) {
-	if len(graphql) == 0 {
-		return nil, ErrEmptyQuery
-	}
-	query, err := parser.ParseQuery(&ast.Source{Input: graphql})
+// writeError writes a GraphQL error response with the given status code and message.
+func (h *Handler) writeError(w http.ResponseWriter, status int, message string, mediaType string) {
+	body, err := json.Marshal(gqlErrorResponse{Errors: []gqlError{{Message: message}}})
 	if err != nil {
-		return nil, err
+		h.logger.Sugar().Errorw("failed to marshal error response", "error", err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
 	}
-
-	rootCollections := make([]string, 0, 1)
-	for _, op := range query.Operations {
-		for _, sel := range op.SelectionSet {
-			if field, ok := sel.(*ast.Field); ok {
-				rootCollections = append(rootCollections, field.Name)
-			}
-		}
+	w.Header().Set("Content-Type", mediaType+"; charset=utf-8")
+	w.WriteHeader(status)
+	_, err = w.Write(body)
+	if err != nil {
+		h.logger.Sugar().Errorw("failed to write error response", "error", err)
 	}
-	return rootCollections, nil
 }
